[level 2] Title: 더 맵게, Time: 1423.11 ms, Memory: 123 MB -BaekjoonHub

+++ 프로그래머스/2/42626. 더 맵게/더 맵게.java
@@ -29,6 +29,6 @@ class Solution {
         return -1;
     }
     public static int getScoville(int a, int b) {
-        return (int) (a + (b * 2));
+        return (a + (b * 2));
     }
 }
